Migracija za User Role, Autentifikacija i autorizacija po rolama.

+++ backend/agencija-backend/routes/api.php
@@ -10,34 +10,34 @@ use Illuminate\Support\Facades\Route;
 
 
 Route::get('/cities', [CityController::class,'index']);
-Route::post('/cities/store', [CityController::class,'store']);
-Route::put('/cities/update', [CityController::class,'update']);
-Route::delete('/cities/destroy', [CityController::class,'destroy']);
+Route::post('/cities/store', [CityController::class,'store'])->middleware(['auth:sanctum', 'ability:user,admin']);
+Route::put('/cities/update', [CityController::class,'update'])->middleware(['auth:sanctum', 'ability:user,admin']);
+Route::delete('/cities/destroy', [CityController::class,'destroy'])->middleware(['auth:sanctum', 'ability:admin']);
 Route::get('/cities/{id}', [CityController::class,'showById']);
 
 
 Route::get('/researchers', [ResearcherController::class,'index']);
-Route::post('/researchers/store', [ResearcherController::class,'store']);
-Route::put('/researchers/update', [ResearcherController::class,'update']);
-Route::delete('/researchers/delete', [ResearcherController::class,'destroy']);
+Route::post('/researchers/store', [ResearcherController::class,'store'])->middleware(['auth:sanctum', 'ability:user,admin']);
+Route::put('/researchers/update', [ResearcherController::class,'update'])->middleware(['auth:sanctum', 'ability:user,admin']);
+Route::delete('/researchers/delete', [ResearcherController::class,'destroy'])->middleware(['auth:sanctum', 'ability:admin']);
 Route::get('/researchers/{id}', [ResearcherController::class,'showByIdSimple']);
 
 Route::get('/publications', [PublicationController::class,'index']);
-Route::post('/publications/store', [PublicationController::class,'store']);
-Route::put('/publications/update', [PublicationController::class,'update']);
-Route::delete('/publications/delete', [PublicationController::class,'destroy']);
+Route::post('/publications/store', [PublicationController::class,'store'])->middleware(['auth:sanctum', 'ability:user,admin']);
+Route::put('/publications/update', [PublicationController::class,'update'])->middleware(['auth:sanctum', 'ability:user,admin']);
+Route::delete('/publications/delete', [PublicationController::class,'destroy'])->middleware(['auth:sanctum', 'ability:admin']);
 Route::get('/publications/{id}', [PublicationController::class,'showById']);
 
 Route::get('/publicationResearchers', [PublicationResearcherController::class,'index']);
-Route::post('/publicationResearchers/store', [PublicationResearcherController::class,'store']);
-Route::delete('/publicationResearchers/delete', [PublicationResearcherController::class,'destroy']);
-Route::delete('/publicationResearchers/delete/{id}', [PublicationResearcherController::class,'destroyById']);
+Route::post('/publicationResearchers/store', [PublicationResearcherController::class,'store'])->middleware(['auth:sanctum', 'ability:user,admin']);
+Route::delete('/publicationResearchers/delete', [PublicationResearcherController::class,'destroy'])->middleware(['auth:sanctum', 'ability:user,admin']);
+Route::delete('/publicationResearchers/delete/{id}', [PublicationResearcherController::class,'destroyById'])->middleware(['auth:sanctum', 'ability:admin']);
 Route::get('/publicationResearchers/{id}', [PublicationResearcherController::class,'showById']);
 
 Route::get('/references', [ReferenceController::class,'index']);
-Route::post('/references/store', [ReferenceController::class,'store']);
-Route::delete('/references/delete', [ReferenceController::class,'destroy']);
-Route::delete('/references/delete/{id}', [ReferenceController::class,'destroyById']);
+Route::post('/references/store', [ReferenceController::class,'store'])->middleware(['auth:sanctum', 'ability:user,admin']);
+Route::delete('/references/delete', [ReferenceController::class,'destroy'])->middleware(['auth:sanctum', 'ability:user,admin']);
+Route::delete('/references/delete/{id}', [ReferenceController::class,'destroyById'])->middleware(['auth:sanctum', 'ability:admin']);
 Route::get('/references/{id}', [ReferenceController::class,'showById']);
 
 Route::post('/register', [AuthController::class,'register']);
